♻️Refactor: DateFormatter의 yesterday프로퍼티를 Date로 이동
